+++ Source/StrategyProj/CharacterInterface.h
@@ -67,11 +67,11 @@ struct FCharacterStatus
 
 	// HP(体力)
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "CharacterStatus")
-		int MaxHP;
+		float MaxHP;
 
 	// SP(スキルポイント)
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "CharacterStatus")
-		int MaxSP;
+		float MaxSP;
 
 	// 攻撃力
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "CharacterStatus")
